PZA-40 Issue with util in scss file
