--- src/Provider/Skeleton/ProductCardSkeleton.tsx
@@ -1,16 +1,0 @@
-import * as React from 'react';
-import Skeleton from '@mui/material/Skeleton';
-import Stack from '@mui/material/Stack';
-
-export default function ProductCardSkeleton() {
-  return (
-    <Stack spacing={1}>
-      <Skeleton
-        variant="rectangular"
-        width={250}
-        height={320}
-        animation="wave"
-      />
-    </Stack>
-  );
-}
