+++ src/assets/icons/index.ts
