+++ src/components/CountDown/CountDownCard.tsx
