--- src/scss/Utils/_drawerProvider.scss
@@ -1,4 +1,0 @@
-.drawer__btn{
-    background-color: white;
-    border: none;
-}
