+++ src/scss/components/_hoverImageContainer.scss
