+++ src/scss/components/_searchBar.scss
