+++ src/scss/components/countdown/_countDownCard.scss
