+++ src/scss/layouts/_collection.scss
