+++ src/scss/pages/_collectionPage.scss
