+++ src/scss/utils/_drawerProvider.scss
@@ -1,0 +1,4 @@
+.drawer__btn{
+    background-color: white;
+    border: none;
+}
